+++ src/main/java/ru/skillbox/socialnet/dto/response/PostRs.java
@@ -12,7 +12,7 @@ public class PostRs {
     private PersonRs author;
     private Set<CommentRs> comments;
     private Long id;
-    private Integer likes;
+    private Long likes;
     private Set<String> tags;
     private String time;
     private String title;
